fixed error when database is not yet configured
