+++ etc/docker/nginx/nginx.conf
@@ -1,26 +1,38 @@
 server {
-    listen 80 default;
-    server_name localhost;
-    client_max_body_size 10008M;
-
-    access_log /var/log/nginx/yawik.access.log;
-
     root /var/www/yawik/public;
-    index index.php;
 
-    if (!-e $request_filename) {
-        rewrite ^.*$ /index.php last;
+    location / {
+        # try to serve file directly, fallback to index.php
+        try_files $uri /index.php$is_args$args;
     }
 
-    location ~ \.php$ {
+    location ~ ^/index\.php(/|$) {
+        # Comment the next line and uncomment the next to enable dynamic resolution (incompatible with Kubernetes)
         fastcgi_pass php:9000;
-        fastcgi_index index.php;
-        fastcgi_param SCRIPT_FILENAME $document_root$fastcgi_script_name;
-        fastcgi_param PHP_VALUE "error_log=/var/log/nginx/yawik_php_errors.log";
-        fastcgi_param APPLICATION_ENV "production";
-        fastcgi_buffers 16 16k;
-        fastcgi_buffer_size 32k;
+        #resolver 127.0.0.11;
+        #set $upstream_host php;
+        #fastcgi_pass $upstream_host:9000;
+
+        fastcgi_split_path_info ^(.+\.php)(/.*)$;
         include fastcgi_params;
+        # When you are using symlinks to link the document root to the
+        # current version of your application, you should pass the real
+        # application path instead of the path to the symlink to PHP
+        # FPM.
+        # Otherwise, PHP's OPcache may not properly detect changes to
+        # your PHP files (see https://github.com/zendtech/ZendOptimizerPlus/issues/126
+        # for more information).
+        fastcgi_param SCRIPT_FILENAME $realpath_root$fastcgi_script_name;
+        fastcgi_param DOCUMENT_ROOT $realpath_root;
+        # Prevents URIs that include the front controller. This will 404:
+        # http://domain.tld/index.php/some-path
+        # Remove the internal directive to allow URIs like this
+        internal;
     }
-    
+
+    # return 404 for all other php files not matching the front controller
+    # this prevents access to other php files you don't want to be accessible.
+    #location ~ \.php$ {
+    #  return 404;
+    #}
 }
