+++ etc/docker/nginx/old.conf
@@ -1,0 +1,28 @@
+server {
+    listen 80 default;
+    server_name localhost;
+    client_max_body_size 10008M;
+
+    fastcgi_param SCRIPT_FILENAME $realpath_root$fastcgi_script_name;
+    fastcgi_param DOCUMENT_ROOT $realpath_root;
+    access_log /var/log/nginx/yawik.access.log;
+
+    root /var/www/yawik/public;
+    index index.php;
+
+    if (!-e $request_filename) {
+        rewrite ^.*$ /index.php last;
+    }
+
+    location ~ \.php$ {
+        fastcgi_pass php:9000;
+        fastcgi_index index.php;
+        fastcgi_param SCRIPT_FILENAME $document_root$fastcgi_script_name;
+        fastcgi_param PHP_VALUE "error_log=/var/log/nginx/yawik_php_errors.log";
+        fastcgi_buffers 16 16k;
+        fastcgi_buffer_size 32k;
+        include fastcgi_params;
+    }
+
+    disable_symlinks off;
+}
