+++ etc/docker/php/Dockerfile
@@ -7,8 +7,6 @@ ENV UMASK 0
 ENV WORKDIR "/var/www/yawik"
 ARG TIMEZONE
 
-ENV http_proxy=http://172.17.0.1:3128
-
 RUN apk upgrade --update && apk --no-cache add \
     gcc g++ make git autoconf tzdata openntpd libcurl curl-dev coreutils \
     libmcrypt-dev freetype-dev libxpm-dev libjpeg-turbo-dev libvpx-dev \
