+++ module/Core/bin/yawik
@@ -21,7 +21,14 @@ if (!class_exists('Core\Module', true)) {
 }
 
 use Core\Console\Application;
-use Core\Bootstrap;
-$mvc = Bootstrap::initApplication();
+use Core\Yawik;
+use Symfony\Component\Dotenv\Dotenv;
+
+if(is_file($file =getcwd().'/.env')){
+    $env = new Dotenv();
+    $env->load($file);
+}
+$mvc = Yawik::initApplication();
+
 $app = new Application($mvc->getServiceManager());
 $app->run();
