+++ public/index.php
@@ -22,6 +22,8 @@ if (file_exists(__DIR__.'/../vendor/autoload.php')) {
     exit;
 }
 
+use Core\Yawik;
+
 chdir(dirname(__DIR__));
 $config = include __DIR__.'/../config/config.php';
-Core\Bootstrap::runApplication($config);
+Yawik::runApplication($config);
